23-06-23_movFeedbackSystem_upload1
First upload of files I've been working with to perform research on the movement feedback system. 
.
I used the Xterity Python application created by Jakob Karolus. Xterity is a recording application (python-based) that discovers LSL streams in the network and allows time-synchronized recording into binary pickle files. 
I've learned openCV and Mediapipe. With this new knowledge, I developed a Python program that recognizes and gives live feedback and guidance on the ideal technical movement the user should perform while shooting a free throw on basketball. The code can be expanded to include different movements in the future.
.
The Xterity application joins all streams (video, sensors, and timestamp markers). It records the streams. Two Xsens DOT movement acceleration sensors are attached to the user's body. One is attached to the wrist, and the other is attached to the hand. The data gathered during the free throw tests are then transformed into a plot of the hand and wrist total acceleration. 
.
The next step in my research is to modify the code so that I can display the Xsens data of hand and wrist total acceleration in real-time to the user. We suspect that the video aid combined with knowledge from the movement acceleration can improve upon the feedback being received and ultimately aid with the process of learning/improving the movement.